refactor : 기존 MainActivity(사진 촬영, 갤러리 사진 선택) -> TakePictureActivity 로 변경, SelectMosaicEditActivity(모자이크, 편집 분기 화면) -> MainActivity로 변경
