--- presentation/src/main/java/team/jsv/icec/ui/fragment/mosaic/SelectFaceFragment.kt
@@ -1,29 +1,0 @@
-package team.jsv.icec.ui.fragment.mosaic
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import team.jsv.presentation.databinding.FragmentSelectFaceBinding
-
-class SelectFaceFragment : Fragment() {
-
-    private var _binding: FragmentSelectFaceBinding? = null
-    private val binding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        _binding = FragmentSelectFaceBinding.inflate(inflater, container, false)
-
-        return binding.root
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        _binding = null
-    }
-}
