+++ presentation/src/main/java/team/jsv/icec/ui/main/MainActivity.kt
@@ -1,31 +1,24 @@
 package team.jsv.icec.ui.main
 
 import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
+import androidx.activity.viewModels
+import dagger.hilt.android.AndroidEntryPoint
+import team.jsv.icec.base.BaseActivity
+import team.jsv.icec.ui.main.mosaic.MosaicViewModel
+import team.jsv.presentation.R
 import team.jsv.presentation.databinding.ActivityMainBinding
 
-class MainActivity : AppCompatActivity() {
-    private lateinit var viewBinding: ActivityMainBinding
+@AndroidEntryPoint
+class MainActivity :
+    BaseActivity<ActivityMainBinding>(R.layout.activity_main) {
+
+    private val viewModel: MosaicViewModel by viewModels()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        viewBinding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(viewBinding.root)
 
-        viewBinding.ratioButton.setOnClickListener {
-            // 비율 옵션 버튼
-        }
-
-        viewBinding.galleryButton.setOnClickListener {
-            // 갤러리 버튼
-        }
-
-        viewBinding.captureButton.setOnClickListener {
-            // 사진 찍기 버튼
-        }
-
-        viewBinding.reverseButton.setOnClickListener {
-            // 좌우 반전 버튼
-        }
+        // TODO(ham2174) : 갤러리 혹은 카메라로 찍은 사진 데이터를 가져와야함. Activity -> Activity data 전달 필요
+        val image = R.drawable.download_icon
+        viewModel.setImage(image)
     }
 }
