+++ presentation/src/main/res/navigation/nav_graph.xml
@@ -6,7 +6,7 @@
 
     <fragment
         android:id="@+id/selectMosaicEditFragment"
-        android:name="team.jsv.icec.ui.fragment.SelectMosaicEditFragment"
+        android:name="team.jsv.icec.ui.main.SelectMosaicEditFragment"
         android:label="SelectMosaicEditFragment" >
         <action
             android:id="@+id/action_selectMosaicEditFragment_to_faceSelectFragment"
@@ -18,12 +18,12 @@
 
     <fragment
         android:id="@+id/faceSelectFragment"
-        android:name="team.jsv.icec.ui.fragment.mosaic.SelectFaceFragment"
+        android:name="team.jsv.icec.ui.main.mosaic.SelectFaceFragment"
         android:label="FaceSelectFragment" />
 
     <fragment
         android:id="@+id/photoEditFragment"
-        android:name="team.jsv.icec.ui.fragment.edit.PhotoEditFragment"
+        android:name="team.jsv.icec.ui.main.edit.PhotoEditFragment"
         android:label="PhotoEditFragment" />
 
 </navigation>
